+++ lib/monitor.js
@@ -10,6 +10,7 @@ var isString = require('101/is-string');
 var isNumber = require('101/is-number');
 var pluck = require('101/pluck');
 
+var Timer = require('./timer');
 var SocketsMonitor = require('./sockets-monitor');
 
 /**
@@ -244,36 +245,4 @@ Monitor.prototype.captureStreamEvents = function (streamName, stream) {
   stream.on('end', sendEvent('end'));
   stream.on('open', sendEvent('open'));
   stream.on('error', sendEvent('error'));
-};
-
-/**
- * Timer class for performing time calculations through the monitor
- * module.
- * @class
- */
-function Timer (callback, start) {
-  this.callback = callback;
-  if (!exists(start) || start !== false) {
-    this.start();
-  }
-}
-
-/**
- * Starts the timer.
- */
-Timer.prototype.start = function () {
-  if (this.startDate) {
-    return;
-  }
-  this.startDate = new Date();
-};
-
-/**
- * Stops the timer and sends information through datadog.
- */
-Timer.prototype.stop = function () {
-  if (!this.startDate) {
-    return;
-  }
-  this.callback(new Date() - this.startDate);
 };
